requirements.txt
# Anime Office Command Center - Dependencies
streamlit>=1.28.0
supabase>=2.0.0
python-dotenv>=1.0.0
